moved search-box to admin only and trying to make the modal dismiss work
