+++ assets/js/admin.ts
@@ -1,3 +1,5 @@
+import "./components/search-box";
+
 import "./admin/codes";
 import "./admin/orders";
 import "./admin/toys_form";
