+++ package.json
@@ -26,13 +26,13 @@
   "homepage": "https://github.com/interactivelabs/reyes-magos-gr#readme",
   "devDependencies": {
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/cli": "^4.1.6",
-    "@tailwindcss/vite": "^4.1.6",
+    "@tailwindcss/cli": "^4.1.7",
+    "@tailwindcss/vite": "^4.1.7",
     "@types/alpinejs": "^3.13.11",
     "esbuild": "^0.25.4",
     "gsap": "^3.13.0",
     "prettier": "^3.5.3",
-    "tailwindcss": "^4.1.6",
+    "tailwindcss": "^4.1.7",
     "tailwindcss-motion": "^1.1.0"
   }
 }
