+++ views/admin/toys/create_toy_form.templ
@@ -13,7 +13,7 @@ templ CreateToyForm() {
 			hx-post={ string(templ.URL("/admin/toys")) }
 			hx-target="#admin-toys-table-body"
 			hx-swap="afterbegin"
-			x-on:htmx:after-load="modalOpen = false"
+			hx-on:htmx:after-load="modalOpen = false"
 		>
 			<div class="grid grid-cols-1 gap-x-6 gap-y-8">
 				@ToyForm(model.Toy{})
