+++ views/admin/toys/update_toy_form.templ
@@ -13,7 +13,7 @@ templ UpdateToyForm(toy model.Toy) {
 			hx-put={ string(templ.URL(fmt.Sprintf("/admin/toys/%d/save", toy.ToyID))) }
 			hx-target={ fmt.Sprintf("#admin-toys-row-%d", toy.ToyID) }
 			hx-swap="outerHTML"
-			x-on:htmx:after-load="modalOpen = false"
+			hx-on:htmx:after-load="modalOpen = false"
 		>
 			<div class="grid grid-cols-1 gap-x-6 gap-y-8">
 				@ToyForm(toy)
