+++ views/admin/volunteers/create_volunteer_form_templ.go
@@ -59,7 +59,7 @@ func CreateVolunteerForm() templ.Component {
 			if templ_7745c5c3_Err != nil {
 				return templ_7745c5c3_Err
 			}
-			templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 2, "\" hx-target=\"#admin-volunteers-table-body\" hx-swap=\"afterbegin\" x-on:htmx:after-load=\"modalOpen = false\"><div class=\"grid grid-cols-1 gap-x-6 gap-y-8\">")
+			templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 2, "\" hx-target=\"#admin-volunteers-table-body\" hx-swap=\"afterbegin\" hx-on:htmx:after-load=\"modalOpen = false\"><div class=\"grid grid-cols-1 gap-x-6 gap-y-8\">")
 			if templ_7745c5c3_Err != nil {
 				return templ_7745c5c3_Err
 			}
