+++ views/admin/volunteers/update_volunteer_form.templ
@@ -13,7 +13,7 @@ templ UpdateVolunteerForm(volunteer model.Volunteer) {
 			hx-put={ string(templ.URL(fmt.Sprintf("/admin/volunteers/%d/save", volunteer.VolunteerID))) }
 			hx-target={ fmt.Sprintf("#admin-volunteer-row-%d", volunteer.VolunteerID) }
 			hx-swap="outerHTML"
-			x-on:htmx:after-load="modalOpen = false"
+			hx-on:htmx:after-load="modalOpen = false"
 		>
 			<div class="grid grid-cols-1 gap-x-6 gap-y-8">
 				@VolunteerForm(volunteer)
